Add toGenerator utility function (#1543)
* Add toGenerator utility function

The toGenerator function is intended for wrapping a promise-returning function
to create a generator-returning function. This can then be used to obtain a
strongly-typed return value in async actions using `yield*`.

* Move toGenerator function to flow module

* Rename toGenerator to toGeneratorFunction

The new name is more accurate, based on the description of generators from MDN:
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Generator

> The Generator object is returned by a generator function

* Add toGenerator to convert a promise to a generator

* Simplify checks for explicit typescript any type

Use a single function that ensures the type passed is of the expected type, and
is NOT the typescript `any` type.