Fix array reconciliation with snapshotPreprocessor
Adds an `isMatchingSnapshotId` method to `ComplexType`, to be used instead of manually comparing identifier values between an existing node and a new snapshot.

Overrides the `isMatchingSnapshotId` method with the `SnapshotProcessor` type to run the preprocessor before attempting to get the identifier value from the new snapshot.

Patches the `is` method on the sub-type of the snapshotProcessor so that the specified snapshot is preprocessed before validation.

Closes #1776