fix: Compatible with TypeScript 3.9 non-enumerable properties
TypeScript 3.9 changes the behavior of [getter/setters so that they are no longer enumerable](https://www.typescriptlang.org/docs/handbook/release-notes/typescript-3-9.html#getterssetters-are-no-longer-enumerable).

The above change breaks mobx-state-tree because it uses `Object.keys(views)` which only return enumerable properties. The correct behavior is to use `Object.getOwnPropertyNames(views)` to retrieve the list of properties to include getter/setters.

This fix updates the code to use `Object.getOwnPropertyNames(views)` and thus be compatible with TypeScript 3.9

Closes #1899