Eliminate `NonEmptyObject`
This was introduced in https://github.com/mobxjs/mobx-state-tree/pull/1269 to
prevent passing in any object to a model type that had no props.

Although this works great, some people have found it harder to work with because
now `keyof typeof MyModelType` now has this extra key they have to deal with.
One could intersect that type with `string`, but it's unfortunate that users
have to do this.

Instead, we use `Record<string, never>` which can only be represented by an
empty object literal. Snapshot preprocessors also had to be tweaked because they
can emit properties that aren't known to them in the case of composition with
other types.