Improve typing for model type substitutability.
Previously, a model derived from an empty model couldn't be used in a place that
asked for the empty model. In general, we had a lot of substitutability issues
(just with our typing).

This commit fixes that by bringing back what was previously known as
"$nonEmptyObject", but improves on it by not having it be as pervasive as it was
previously (for example, it would show up on models that had props).