Fix model types when extending in any way. (#2218)
Previously, we used to have TS resolve the extensions by just doing something
like `A & B`. The problem is that in TypeScript that means "it's both A and B",
but what happens if you have a required prop in one and an optional prop in the
other? Well, the least common denominator wins, which is requiredness, because
it satisfies both the required and optional properties.

MST doesn't work that way though. Whenever we extend a model we "clobber" the
old property of the same name. To model this in TS we need to omit properties
from the previous version of the model and replace them with the newer ones.